Added PathfinderComposition with Connection modifications.
