+++ ts/masters/modules/gantt.src.ts
@@ -13,6 +13,7 @@
 import Highcharts from '../../Core/Globals.js';
 import Connection from '../../Gantt/Connection.js';
 import Navigator from '../../Stock/Navigator/Navigator.js';
+import Pathfinder from '../../Gantt/Pathfinder.js';
 import Scrollbar from '../../Stock/Scrollbar/Scrollbar.js';
 import RangeSelector from '../../Stock/RangeSelector/RangeSelector.js';
 import XRangeSeries from '../../Series/XRange/XRangeSeries.js';
@@ -31,6 +32,7 @@ G.Scrollbar = Scrollbar;
 // Compositions
 ArrowSymbols.compose(G.SVGRenderer);
 CurrentDateIndication.compose(G.Axis, G.PlotLineOrBand);
+Pathfinder.compose(G.Point);
 Navigator.compose(G.Axis, G.Chart, G.Series);
 RangeSelector.compose(G.Axis, G.Chart);
 Scrollbar.compose(G.Axis);
